// part9/flight-diary/frontend/src/types.tsx
/* export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Stormy = 'stormy',
  Windy = 'windy',
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}
 */
export type Entry = {
  id: number;
  date: string;
  weather: string;
  visibility: string;
  comment?: string;
}

export type NewEntry = Omit<Entry, 'id'>;

//export type NonSensitiveEntry = Omit<Entry, 'comment'>;